requirements.txt
asgiref==3.2.7
autopep8==1.5.2
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
pycodestyle==2.5.0
pytz==2019.3
sqlparse==0.3.1
